+++ src/main/java/gov/ca/cwds/healthcheck/CognitoHealthContributor.java
@@ -1,10 +1,11 @@
-package gov.ca.cwds.management;
+package gov.ca.cwds.healthcheck;
 
 import com.amazonaws.services.cognitoidp.model.AWSCognitoIdentityProviderException;
 import gov.ca.cwds.idm.service.CognitoServiceFacade;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.actuate.health.Health;
 import org.springframework.boot.actuate.health.HealthIndicator;
+import org.springframework.boot.actuate.health.Status;
 import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Component;
 
@@ -18,7 +19,7 @@ public class CognitoHealthContributor implements HealthIndicator {
   public Health health() {
     try {
       cognitoServiceFacade.healthCheck();
-      return Health.up().status("Cognito Connection OK").build();
+      return Health.status(new Status("UP", "Connection is OK")).build();
     } catch (AWSCognitoIdentityProviderException awsException) {
       return Health.down().withException(awsException).build();
     }
