--- src/main/resources/version.properties
@@ -1,2 +1,0 @@
-build.version=@build.version@
-build.number=@build.number@
